+++ rc/pksrc.h
@@ -7,6 +7,7 @@
 #define IDD_STEP                        3
 #define IDD_ALL_OCCURRENCES             4
 #define IDD_CONTINUE                    4
+#define IDS_FONT_AWESOME                4
 #define IDS_STRING15                    15
 #define IDS_NO_TEMPLATES_DEFINED        15
 #define IDD_RO6                         80
@@ -45,7 +46,7 @@
 #define IDS_TT_SORT_FIELD_SEPARATOR     205
 #define IDS_TT_SORT_FIELD_OPTIONS       206
 #define IDS_WARN_FILE_MOVED_OR_DELETED  207
-#define IDS_FONT_AWESOME                224
+#define IDR_FONT1                       232
 #define DLGCURSOROPTIONS                529
 #define DLGFINDINFILES                  535
 #define DLG_OPT_WARNINGS                556
@@ -85,7 +86,7 @@
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        228
+#define _APS_NEXT_RESOURCE_VALUE        233
 #define _APS_NEXT_COMMAND_VALUE         40038
 #define _APS_NEXT_CONTROL_VALUE         1018
 #define _APS_NEXT_SYMED_VALUE           101