+++ doc/release_notes.md
@@ -16,6 +16,7 @@
   types.
 - Zoom plus and zoom minus icons have been added to the toolbar. Action Bindings to Ctrl++ and Ctrl+- were also added. Before
   one needed to use the num block for zooming.
+- Icons were updated to FontAwesome Version 6 introducing new icons.
 
 ### Fixes
 - When opening multiple files so that not all file tabs are shown having a tab being selected, which needs to be scrolled into view, this